Major code cleanup

+++ src/main/java/ruukas/infinityeditor/gui/GuiAttributes.java
@@ -272,7 +272,7 @@ public class GuiAttributes extends GuiInfinity {
         
         for ( int i = 0 ; i < sharedAttributes.length ; i++ )
         {
-            double angleI = (((double) (rotOff + (double) (Math.abs( mouseDist - r ) >= 16 ? partialTicks : 0d)) / 60d)) + (angle * i);
+            double angleI = (((rotOff + (Math.abs( mouseDist - r ) >= 16 ? partialTicks : 0d)) / 60d)) + (angle * i);
             int x = (int) (midX + (r * Math.cos( angleI )));
             int y = (int) (midY + (r * Math.sin( angleI )));
